### This pullrequest changes

Instrumentation framework for OpenCV performance test system. It was developed to:
- track implementations execution inside OpenCV functions with call tree output to test logs
- track implementations execution time weight in the pipeline

Framework can be enabled with ENABLE_INSTRUMENTATION define in CMake and disabled by default. In disabled state it doesn't consume resources.

Example instrumentation was implemented for cv::meanStdDev function:

Output example for IPP:
```
[ RUN      ] Size_MatType_meanStdDev.meanStdDev/6
[ PERFSTAT ]    (samples = 25, mean = 0.10, median = 0.10, stddev = 0.00 (0.5%))
[ VALUE    ]    (1920x1080, 8UC1)
[ TRACE    ]
ROOT
\---cv::meanStdDev - C:25 /*entries counter*/ T:2.3906ms /*block execution time*/
    \---cv::ipp_meanStdDev<W_IPP> - C:25 T:2.3269ms L:97% /*weight in local block*/ G:97% /*weight in entire pipeline*/
        \---ippiMean_StdDev_C1R<F_IPP> - C:25 T:2.2700ms L:98% G:95%

IPP weight: 95.0% /*total implementation functions execution weight across all entries*/
OPENCL weight: 0.0%
[/TRACE    ]
[       OK ] Size_MatType_meanStdDev.meanStdDev/6 (77 ms)
```

Output example for OpenCL:
```
[ RUN      ] OCL_MeanStdDevFixture_MeanStdDev.MeanStdDev/7
[ PERFSTAT ]    (samples = 25, mean = 0.23, median = 0.23, stddev = 0.00 (1.8%))
[ VALUE    ]    (1280x720, 32FC1)
[ TRACE    ]
ROOT
\---cv::meanStdDev - C:25 T:5.8007ms
    \---cv::ocl_meanStdDev<W_OCL> - C:25 T:5.7412ms L:99% G:99%
        \---k.run<F_OCL> - C:25 T:3.5630ms L:62% G:61%

IPP weight: 0.0%
OPENCL weight: 61.4%
[/TRACE    ]
[       OK ] OCL_MeanStdDevFixture_MeanStdDev.MeanStdDev/7 (36 ms)
```