@alalek 2016-08-12
Awesome useful feature!