@lupustr3 2016-08-12
"--perf_instrument" command line switch was added to ps