@lupustr3 2016-08-16
- Parallel regions handling was redesigned and improved.
- Total time accumulators were replaced with running mean time accumulators to decrease overhead of threads timing synchronization in parallel regions
- Changes according to notes above.