@alalek 2016-08-16
Great update! Looks good.

BTW, What do you think about [these](https://github.com/opencv/opencv/pull/6069/commits/82762f3b8599e0d6e3c46544c9ae145ee0ea586b) macro changes? Are there any expected problems?