@lupustr3 2016-08-17
I think it should work the same.