@lupustr3 2016-08-17
New macro interface